Break out files, style up account info.
