+++ common/assets/styles/etherwallet-custom.less
@@ -209,38 +209,6 @@ textarea {
   }
 }
 
-.account-info {
-  .clearfix;
-  padding-left: 1em;
-  margin: 0;
-  li {
-    margin-bottom: 0;
-    list-style-type: none;
-    word-break: break-all;
-  }
-  table& {
-    font-weight: 200;
-    border-bottom: 0;
-    min-width: 200px;
-    td {
-      padding: 4px 5px;
-      line-height: 1;
-    }
-    td:first-child {
-      max-width: 115px;
-      word-wrap: break-word;
-      padding-left: 1em;
-    }
-    tr:nth-child(even) {
-      background-color: @gray-lightest;
-    }
-    tr:nth-last-child(2),
-    tr:last-child {
-      background-color: white !important;
-    }
-  }
-}
-
 input[type="text"] + .eye {
   cursor: pointer;
   &:before {
