+++ common/components/BalanceSidebar/Promos.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default class Promos extends React.Component {
+  render() {
+    return <div />;
+  }
+}
