--- common/components/BalanceSidebar/index.js
@@ -1,3 +1,0 @@
-// @flow
-
-export { default } from './BalanceSidebar';
